LibWeb: Make apply_the_history_step return value via a continuation
This is a first step towards making this function run more async.
It no longer has a "normal" return value, but instead invokes a
GC-allocated continuation function when exiting.

The return value was only actually used in the Navigation API, where it
was then used asynchronously to resolve/reject promises, so it fit
naturally with this style.