LibJS: Remove ExecutionContext::this_value and always use register slot
For bytecode functions, we kept the `this` value in two places, both in
ExecutionContext::this_value, and in reg(Register::this_value()).

This was obviously redundant, so let's consolidate them, using only
the register slot. This way bytecode keeps working as before, and we
can simply remove some awkward initialization code.

Note that we also change the this_value index to be register #1. This is
to minimize the amount of registers that NativeFunction needs to
allocate. #0 would be ideal, but register #0 is used in some hard-coded
ways that we'd need to deal with for that to work.