LibJS: Rewrite `i++` to `++i` in common cases where result is ignored
This commit introduces the "ignored" operand, which can be passed as
the preferred_dst to generate_bytecode(). If we end up trying to emit
a postfix increment with "ignored" destination, it now emits a prefix
increment instead.

Turns out postfix increment is quite common on the web, and this avoids
a bit of extra work to preserve the original value for no reason.